Move FeatureMatcherCache to feature/matcher
